checks/menu-format + Drop menu-icon-has-bad-colors, it doesn't apply anymore (Closes: #199341)
I dropped the whole color-parsing bit, as it solely was meant to get the
colors. However, in the process it could have caused an parse error, which is
now not checked for anymore. As the parser is very limited, and would only
find very specific cases, I don't think it's worthwhile to retain it. A such
malformed .xpm is almost not possible unintentionally.