Check for unnecessary versioned dependencies on Perl core modules. (Closes: #497887)
Add a warning about unnecessary versioned dependencies on packages
also provided by the perl core packages at a recent enough version.

In the case of alternate dependencies, only the first alternative is
examined. This allows the recommended idiom of specifying the perl core
package as the preferred alternative and the versioned dependency on
the real package as a secondary one.