Incorporate static data about packages in the Perl core
The new file, data/fields/perl-provides, contains information on
modules included in the Perl core packages that also have a separate
binary package available.

The 'refresh-perl-provides' target in debian/rules provides a way to
semi-automatically refresh the data, and the 'build' target issues
a warning if the data is clearly stale.