debian/changelog: Provide a summary of added/removed tags
I think this is a useful thing to have to be able to quickly determine
how the number of tags changes and which release introduced/dropped
which tags.

I know this can also be extracted from the git log but it can't hurt
to make this easily accessible for all interested persons.