Don't output a "start package" message for a .changes files.
Don't output a "Processing binary package foo.changes (version )" message.
Thanks Sandro Tosi for pointing the issue out.

.changes files aren't really packages, but we pretend they're binary
packages so that we can issue tags against them.

We already output a message indicating we're processing a .changes
file so the "start package" message is spurious and somewhat confusing.
Maybe we should have a "changes" type rather than faking them as binary
packages, but in the meantime this gets us back to the pre-::Output
output.