Parameterize the section of test suite packages and clean up changelogs
Allow the section of a package generated by the test suite to be set in
the *.desc file.  This makes it easier to write tests for checks on some
packages without introducing spurious tags.

Remove extra whitespace in debian/changelog to group by root directory.

Collapse the changelog for the test suite skel package to only increment
the version once since the previous Lintian release.