Move private/tags-never-seen to t/COVERAGE and add legacy report
Rename private/update-never-seen to private/update-coverage and modify
it to generate t/COVERAGE directly instead of printing to standard
output.  Move private/tags-never-seen to t/COVERAGE.  Add a list of all
tags only tested by the legacy test suite to t/COVERAGE so that we can
track our progress in converting to the new test suite.