Add reporting harness option to disable signature checks
Add an option to harness to "disable" the gpg check performed by
dpkg-source prior to unpacking by pretending that /bin/true is gpg.

The signature check was useless because the required gnupg setup was
not available.

This reduces time when unpacking and size in the lintian.log file, by
assuming that the signatures are fine, as the packages would otherwise not
be in the archive in the first place.

Signed-off-by: Raphael Geissert <atomo64@gmail.com>