Improvements to the duplicate font check
Use space as a delimiter instead of an equal sign as it's a bit easier
to read.  (For future work, we could align columns.)

Fix some minor coding style issues in checks/files.

Add more documentation and a file header to the data file in
private/refresh-fonts-data and handle the case of multiple files providing
the same font by dropping the second package.  (Currently, this only
occurs once in the archive and the second package is non-free.)  Fix a
few minor style issues.

Fix the test to properly declare what tags it checks for.