Further fixes to private/refresh-fonts-data
Honor the contents file as a second argument.  Don't cd into the working
directory and instead qualify paths.  Fix some problems I introduced in
the regex used to process the contents file and really handle the case
of multiple packages.  Atomically replace the data file with the new
version rather than overwriting it.