Don't warn about non-numeric compat levels if no compat was specified
Use an undefined value as a marker that we haven't seen any compat level
at all and don't warn about non-numeric compat levels in that case.
Also rename $needversiondepends to $level, since it's really being used
to hold the configured debhelper compatibility level.  (The variable is
also shorter and easier to type.)