Dynamically generate the list of virtual packages
Replace the hard-coded list of virtual packages with one that can be built
with a script, which lists virtual packages provided by two or more
packages.

This eases the maintenance of such list by knowing about new packages and
dropping obsolete packages.

Since as of the time of writing not all the packages that used to be on
the hard-coded list are provided by more than one package, their inclusion
on the list can be forced as long as there is at least one package
providing it on the archive.