Fix a logic error checking the position of meta-information in symbols files.
When parsing symbols files, correctly ensure that meta-information occurs
between the end of the dependency template(s) and the start of the symbol
list.  The code that was intended to check this resulted in a false positive
syntax error being generated if a file with alternative dependency templates
used meta-information.