Report truncated binaries separate from corrupted binaries
Catch "file truncated" errors from objdump and warn about them.  A version
of binutils recently (as of the time of writting) uploaded to sid,
correctly, started to emit this error instead of just saying the binary
was corrupted.

This affected the binary-from-other-arch test since it only uses the
header of an ELF object.