Changelog and tweaks for init script checks
Lower init.d-script-uses-usr-interpreter to normal/possible.  This is not
a Policy violation (even of a should) and it's possible to do safely.
It's just normally not a good idea.

Ignore /etc/init.d/skeleton in init script checks as well.

Rename checkinit to check_init following naming conventions elsewhere.
We'll change globally if we start treating check_* subs as special.

Make Emacs happier with comment markers in Perl regexes.