Handle string collection from set[gu]id files when run as root
* collection/strings:
  + [ADB] Handle the fact that, when Lintian is run as root, the output of
    "file" on set[gu]id files may include the fact that they are set[gu]id.
    Thanks, Peter Pentchev.  (Closes: #533618)

When run as root, the files in the unpacked directory may be created with
set[gu]id bits set if the bits are set in the binary package.  The sed
expression extracting ELF binaries from the file-info file therefore needs
to handle the differing output in this case.

This isn't an issue when the user is not root, as the unpacked files will
not have the bits set.