Fix missing-separator-between-items' handling of "foo(>= 3)"
If there was no space between the package name and the version relation,
missing-separator-between-items was being incorrectly raised for
"foo(>=" and "3)".  dpkg normalizes the relation to include a space so
Lintian should handle such cases.

Also simplify the regex a little now that it no longer needs to handle
architecture lists.