Fix false positives for watch file mangling
The debian-watch-file-should-mangle-version check was not taking
whether an upstream version was specified in the watch file or not,
therefore producing false positives.

Additionally, the check was being run on a global scope, even when
multiple websites were being specified. This has been fixed and made
those checks indicate the offending lines.