Recommend doc-base registration if PDF or HTML docs found
HTML and pdf files are usually documentation and should therefore be
registered in doc-base.

The check emits the tag only if no doc-base registration file is
provided in the package (instead of trying to look for them there) to
reduce the number of false positives. The theory behind this is that if
the maintainer is already shipping a doc-base registration file she
knows what she is doing.