Tweak check for doc-base registration to exclude changelog and README
Exclude changelog.html.gz from requiring doc-base registration since it
has a special place in Policy and may not be regular documentation.
Also exclude README files in HTML -- this is iffier, and maybe they
should be registered, but a lot of them are small and insignificant and
I suspect maintainers aren't going to want to register all of them.