Move tag statistics code to html_reports
This is the first step to integrate graph generation into html_reports. Both
tags.tmpl and tags-severity.tmpl included code to get tag-specific statistics
(number of tags, overrides and packages). Even though it is simple, we need to
do it again in html_reports anyway in order to keep the record of previous
values. Removing the duplicate code altogether seems like a good idea.

* reporting/html_reports:
  + [JP] Accumulate tag-specific statistics in a new variable, and pass it
    to the appropriate templates.
* reporting/templates/tags{,-severity}.tmpl:
  + [JP] Remove duplicate code to calculate tag statistics.