Fix issues with lintian-clean reports
The maintainer's id was missing from the list of variables passed to the
templates, so the displayed URLs to the full report were wrong. However, for
fully lintian-clean reports it doesn't make sense to link to the full report
since it doesn't provide additional information. Those links have now been
removed from lintian-clean reports.

Also, note that the maintainer's id is now passed to the "clean" template for
completeness' sake. It isn't strictly needed, but it may be useful in the
future.

* reporting/{html_reports,templates/clean.tmpl}:
  + [JP] Don't link to full report if the maintainer only has clean
    packages. (Closes: #545781)