First draft of new Lintian::Tags module
This module is intended to replace Tags.pm.  It includes all the
functionality of that module that was actually used, except for some
parts that moved to Lintian::Tag::Info.  It also has the code to set
display levels based on severity and certainty, which will eventually
replace the code currently in frontend/lintian.

This is not a drop-in replacement.  Adjustments to other code to use
it will be done in subsequent commits.