Rework private/transtats to just show tag statistics
The translation of all tags into severity/certainty is done, so we don't
need the part that checks that, but it's still useful to have statistics
for tag classification.  Rename private/transtats to private/tag-stats
and retain only those portions.  Fix a few bugs in the process and sort
the tag and collection lists.