Optimise spell-checking code by dropping call to length()
Tightening the regex to only match {1,5} uppercase characters instead
of first checking the length of the string and then matching any number
of upper case characters is faster.

This optimisation has an impact on words where the first character is
not uppercase and those with more than five uppercase characters at the
beginning of the word.