Don't warn about length of short description of text templates
For udebs all cases are templates of type "text" that contain text
fragments that get dynamically included in other templates at
runtime. Because of that specific usage, the length of the short
description is not relevant.

To avoid these false positives, please add an exception for this
warning:
- if the template type is "text" AND
- the package is a udeb

Reasons to request the exception are:
- udebs cannot contain lintian exceptions for binary packages (because
  we don't want them installed in the D-I environment)
- we *do* want the warning for other template types than "text"

Also change the name of the local variable (template) 'type' which
currently shadows the higher level (package) 'type' variable.