checks/debian-source-dir: add check for exporting from d/s/git-patches
If debian/source/git-patches exists, read debian/patches/series, and
see if the first line matches a comment emitted by the relevant gitpkg
hook.

Rationale: version 0.17 of gitpkg adds the ability to export arbitrary
ranges of git commits as quilt patches. This exporting is optional,
and this check is intended to warn a user who accidentaly fails to
enable it.