Exclude package names from regular spell checks
Lintian::Check::spell_check now accepts a hashref as the last
argument.  Any key in this hashref will not be considered a
spelling error.  All checks using it has been changed to pass
the name of the current package as an exception.

Note that these exceptions are ignored for multi word spell
checks and the "picky" (case sensitive) spell checks.

Lintian/Check.pm
    + [NT] Allow package names to be excluded from regular spelling
      checks.  They are still subject to multi word and case sensitive
      spelling checks.  Thanks to syq for the suggestion.
      (Closes: #576283)