Replaced unpack-changes-l1
unpack-changes-l1 has been replaced by collection scripts and a little
perl sub (in Lab::Package) to create a the base dir and symlink the
changes file.  The old unpack() sub has been renamed to _unpack() and
now is called by the new create_entry() if Lintian still needs the
unpack script for the given $pkg_type.