Check return value of close and unlink in collections
A number of collections did not check the return value of close
on a write handle.  This could lead a collection to succeed even
if it failed to collect the data it was supposed to.  Return value
of close on a read handle is still (usually) ignored.
  The same thing was done for unlink to ensure that files/symlinks
are really unlinked.