Assume that any of /,&,|,;,$,",',< and > are malicious
They are not allowed in the package/source/version/arch fields
and they are usually interpretted by the shell.  Since perl
open allow magic like "cmd1 | cmd2" and a lot of different
code styles are spread through out lintian, there may be
an exploitable window for some of these characters.