Merge branch 'vendor-profile' into master
This (squashed) commit adds Vendor Profile support for
Lintian. lintian and lintian-info now supports the option
 --profile <vendor/profile>

In the absence of that (or --ftp-master-rejects for
lintian), they will fallback to using dpkg-vendor
(or Dpkg::Vendor) to find the first related vendor
that has a profile and use that.

The full specification of the Vendor Profiles have been
added to the Lintian User Manual.