Added support for multi-version/arch in temporary labs
This patch makes temporary labs able to extract and maintain more
than one instance of the same package as long as the version
or/and the architecture combinations are unique.  This allows for
things like:

  $ lintian -v pkg_1.0-{1,2}_amd64.deb pkg_1.0-{1,2}_i386.deb

without causing an internal error (see #632115).

The LAB_FORMAT has not been bumped since temporary labs are not to
be reused anyway and static labs do not have this feature enabled.

The emitted tag output remain unchanged, since some may depend on
the output format.  Thus this feature is rather hard to use witout
--verbose.