Refactored Read_pkglists and the parts using it
Read_pkglists no longer relies on global variables, in which to store
the results.  Instead, it will generate a new hash each time it is
called and return the ref to it.

This allows for the removal of a couple of:

 # no clue why this is here
 use vars qw(...) # from the above

It also no longer falls back to the info files in the Lab, so files
have to be passed explicitly now.  Nor does it accept the $quiet
parameter anymore that as a side effect completely changed the
error handling.

This means that frontend/lintian will now terminate if one of the
files cannot be read instead of continuing and failing with a
"Cannot find package XYZ in dist or lab".

The list-{bin,src,udeb}pkg scripts will happily ignore the contents of
the packages file if it cannot be read or the format is not
compatiable.  If only because this is what they used to do...