Always produce a tag page, even if it was not emitted
html_reports will now always generate a page for tags, even if
they were not seen in the log.  The pages will mention that the
tag was not seen in the run.  There is also a sorted list of all
tags known to Lintian.

This allows people (and other services) to "blindly" refer to the
http://lintian.debian.org/tags/$tag.html for a description of the
tag.