use Dpkg::Version in checks/fields to validate and split versions
Dpkg::Version does a better job with validating some versions[1] and
is already available due to the libdpkg-perl (runtime) dependency.
The Build-Depends have been updated to include libdpkg-perl,
previously it was not required since libdpkg-perl was only used as
a fallback when dpkg-dev (build-essential) was not present.

[1] Cases including invalid epochs (e.g. a:0-1), which the old
Lintian version checker would "accept" by considering the epoch
for a part of the upstream version.