Remove the respect-display-level code
The resulting semantics could trivially be achieved with a few
calls to other parts of the Tags API.  Furthermore the code made
Lintian::Tags harder to understand.

This patch changes the semantics of --check-part so it now uses
display-level settings again.  This is more consistent than the
"intended 2.5.2" behaviour where --display-info was assumed.

The --tags and --tags-from-file options will display the tags
regardless of the display setting.  If the user has explicitly
requested the tag, he/she should not have to pass up to 3
other options to see it.