Improve the behaviour of --quiet and --debug/--verbose
If --quiet is used to --debug, --quiet is silently ignored and
--verbose is unconditionally enabled.  If --quiet and --verbose are
both present (and --debug is not used), the last of the two options
will take effect.

The frontend now accepts --quiet in the lintianrc via the new "quiet"
variable.  The lintianrc.example has been updated to include an
example.

The manpage has also been updated to describe this behaviour.  It also
clarifies that --verbose can be used in the lintianrc file.  The
frontend accepted this in 2.5.2, but the manpage did not mention this.

In Lintian::Output "verbose" and "quiet" has been replaced by
"verbosity_level".  The old "quiet" is a verbosity_level of -1 and
"verbose" has become a verbosity_level of 1.