Skip possible upstream changelogs in coll/changelog-file
coll/changelog-file now double checks changelogs that does not
have "debian" in their filename.  If such a changelog fails the
check, it is assumed to be an upstream changelog and the Debian
one is assumed to be missing.

This avoids a myriad of "syntax errors" tags in packages like
perl, where the Debian changelog is missing and the upstream
one is present.