Migrated reporting/html_reports to Lintian::Lab::Manifest
It has an implicit assumption that there is only one entry of each
type with the same name.  For now it just picks one of them (as
determined by visit_all).  It should not be a huge issue as long as
lintian.d.o-like setups do not support multiple architectures.

Signed-off-by: Niels Thykier <niels@thykier.net>