Use the source name as name of changes processables
Previously the "name" of a changes processable was the base name
without the ".changes" extension.  However, this "adds" the version
and the (build host) architecture to the name[1].

The consequence of this was that a changes file is stored in a
different part of the lab than source (and binary) packages.  Example:

 l/lintian/             - binary and source are stored here
 l/lintian_2.5.2_amd64/ - changes file is stored here

By using the source field[2], the changes are now stored next to the
relevant source package.

This also makes querying the lab for a changes file a lot easier.  Compare
the old query:

  $lab->get_package ("lintian_2.5.2_amd64", '2.5.2', 'source all');

to the new query:

  $lab->get_package ('lintian', '2.5.2', 'source all');

[1] The latter being different from $proc->pkg_arch in a lot of cases.

[2] In its absence, fall back to the first part of the base name
excluding "_".

Signed-off-by: Niels Thykier <niels@thykier.net>