Remove dist search and create a new lab query format
The lab query format allows more control over which package is
requested instead of "all with the name foo".  The query format
is:

 [type:]package[/version[/arch]]

The "type" is one of "ALL", "binary", "source", "udeb" and "changes".
"package" is the name of the package (as before).  "version" is the
exact version of the package requested and "arch" is the specific
architecture of the package.

If version or arch is left out, they are assumed to have the specical
value "_".  If version or arch is "_", then the parameter is ignored
(i.e. "any version" or "any architecture").  The "arch" parameter
is completely ignored when matching source packages.

Signed-off-by: Niels Thykier <niels@thykier.net>