Create lab-info file for labs to store important meta data
The lab-info file stores the Lab format and also a new "Layout"
parameter.  The layout parameter decides how packages are stored
in the lab (currently only "pool" is accepted).

This "trivially" allows testing of different layout formats without
having to change the lab-format.  Implementation-wise this should be
as easy as writing a sub like "_pool_path" and calling that instead
for the different layout.

Signed-off-by: Niels Thykier <niels@thykier.net>