Added experimental output that always uses full package info
The new output format is basically the old output format, except it
always uses the full metadata (even if it is redundant).  As an
example

 T: my-pkg udeb: tag

becomes

 T: my-pkg udeb (1.0-2) [i386]: tag

This allows harness script (and other machine-parsers) to always get
the full package information with the tag without using state-based
parsing (i.e. checking for "N: Processing xyz ..." lines).
  It also disambiguates which package is being process without needing
--verbose.  Optimally the normal output would add the extra fields as
needed, but currently it does not.

On the downside this format is a bit terse for human-readers and is
probably not a good candidate as a default (interactive) output.

Signed-off-by: Niels Thykier <niels@thykier.net>