L::Lab: Fixed 2 bugs and changed get_package API
Adding missing use of Lintian::Collect and error out in _get_lab_index
if the type is undefined.

Updated the get_package documentation to reflect that the entry can
only be created if a Lintian::Processable is used.  The other method
would silently fail and this was the "simplest" solution.

Signed-off-by: Niels Thykier <niels@thykier.net>