Weed out a set of bugs breaking harness
Pass an instance of L::Proccessable::Package to $LAB->get_package (see
commit 71ff473), when adding a nw package.  Fixed the argument order
for $LAB->get_package when removing a package.

Fixed some method calls to non-existent methods.

Remember to actually parse the Sources file and explicitly close the
handle.  Add the package timestamp to the entries as the diff uses
this to see if a package has changed.

Signed-off-by: Niels Thykier <niels@thykier.net>