Fixed an issue where auto-clean would not see the true coll status
Since 2.5.3 the processables may be instances of L::Lab::Entry
themselves.  But in unpack the Lab would be unconditionally queried
for a new entry even if the proc was an entry.

This new entry would be used for recording coll statuses, but the
original entry would be sent to auto-cleaning.  As the collection
status for the original entry was not updated, auto-clean would assume
there was nothing to do and left unpacked on the disk.

Signed-off-by: Niels Thykier <niels@thykier.net>