Corrected any-CPU wildcards for some archs
This is done by mapping architecture names to their "DEB_HOST_ARCH_OS"
and "DEB_HOST_ARCH_CPU" values and using those to determine the
wildcards.

Previously architecture wildcards were determined from the Debian
arch by splitting the architecture by "-" and using the result to
define "OS-any" and "any-CPU".

This relies on the assumption that the CPU wildcard has the exact same
CPU name as the Debian arch.  Unfortunately, this assumption is
invalid for some cases (e.g. armel and lpia)[1].

[1] This can be confirmed by using dpkg-architecture:

 $ dpkg-architecture -aarmel -iany-armel && echo yes
 $ dpkg-architecture -aarmel -iany-arm && echo yes
 yes

 $ dpkg-architecture -alpia -iany-lpia && echo yes
 $ dpkg-architecture -alpia -iany-i386 && echo yes
 yes

Signed-off-by: Niels Thykier <niels@thykier.net>