L::Profile: Improve the "recursive definition"-check
Catch profiles that are recursively defined, but where the Profile
field of the child and Extend field of the parent are not identical.
This can happen if one profile uses "$vendor" and another uses
"$vendor/main" or if symlinks are involved.

This patch solves the issue by deferring the check until the child
profile has been read.  Then it can use the Profile field as the
"canonical" name for the profile.

Note the parents field will use the value in the Extends field and not
the Profile field in the child profile.

Signed-off-by: Niels Thykier <niels@thykier.net>