Removed Checker module and reduced the "magic" of LINTIAN_ROOT
The code from the Checker module has been inlined into its own sub in
frontend/lintian.  The extra module did not reduce the size of
f/lintian notabily and it allowed a reduction in the ways we propagate
LINTIAN_ROOT to modules.

Previously LINTIAN_ROOT needed to be exported as an ENV variable AND
as an "our $LINTIAN_ROOT".  Now, only the ENV variable is needed.

Signed-off-by: Niels Thykier <niels@thykier.net>