Lintian::Lab: Added repair_lab method
This is able to correct lab indices referring to entries that are
either broken or does not exist.  The solution is to remove the
metadata reference and the remains of the entry (if any).

It lacks a check for packages present in the lab that does not appear
in the relevant index.

Signed-off-by: Niels Thykier <niels@thykier.net>