L::Data: Pass CURVALUE argument to CODE when loading data files
Pass a new third argument to the CODE parameter when loading a data
file.  The third argument (CURVALUE) is the current value associated
with a given key (if any).  The CODE sub can modify CURVALUE to built
more complex data structures from multiple lines.

The CODE sub may now (only) return undef if CURVALUE is defined.  In
this case, CURVALUE remains the value associated with the KEY.  This
avoids needless creation of new values when the original value can be
modified instead.

With these changes it is possible to parse output/manual-references
via L::Data instead of the ad-hoc parser currently used in
L::Tag::Info.

Signed-off-by: Niels Thykier <niels@thykier.net>