L::Relation: Add "and" constructor
Exploit that L::Relation objects are immutable to create an "and"
constructor.  This constructor creats an L::Relation object from
L::Relation objects and textual relations. The new L::Relation object
will use the internal structure of other L::Relation objects were
possible to reduce overhead.

The main advantage of this constructor is that it allows better reuse
of the L::Collect::relation method

Signed-off-by: Niels Thykier <niels@thykier.net>