L::Relation: Add empty() and matches(regex, rule) methods
Add method to match predicates or entire OR clauses against a regex
it.  This is mostly useful when checking if a "pattern" package is in
the dependency (e.g. phpapi-$ver).  Generally implies(_inverse) is a
better method, but it does not work if there is not an exact package
name.

Signed-off-by: Niels Thykier <niels@thykier.net>