coll/objdump-info: dctrl-ify the output
Instead of relying on making "readelf output look like objdump",
filter out the unneeded part and write the interesting data as
a dctrl formatted file.

As a side effect, merge apparently-truncated-elf-binary into
apparently-corrupted-elf-binary.  Unlike objdump, readelf does not
really give a reliable way of distinguishing the two cases.

The patch also fixes a couple of references to objdump in the
tag description of two tags.

Signed-off-by: Niels Thykier <niels@thykier.net>