p/refresh-manual-refs: report all missing manuals
Report all missing manuals (rather than only the first).  This is
prevents a tedious "install + refresh" cycle when most of the manuals
are not installed on the system.

Also print the "index" file for the manual.  This makes it easier to
use (e.g.) apt-file to find the manual when the manual name is not a
1:1 match with the package name.

Signed-off-by: Niels Thykier <niels@thykier.net>