L::Util: Remove Lintian::Output dependency
This does cause a slight behavioural change to "fail()".  Previously
fail would join multiple arguments with newline (and prefix each line
with "internal error: "); now it joins arugments with a space.
  The multiple argument calling convention does not seem to be used
currently, so this change should not cause any issues.

The second change is that arugment-less calls to fail will now include
filename and line number of the caller.  This gives slightly better
possibility for tracking down the issue - compare:
  internal error: no such file or direction
 - to -
  internal error: no such file or direction (called from $filename:$line)

Signed-off-by: Niels Thykier <niels@thykier.net>