lintian: Fix race condition when reaping jobs
In some cases, Lintian's selection of jobs caused it to skip some
unpacking jobs.  If all the jobs currently in the queue would not add
any new jobs, Lintian would assume it was done.  This turns out to be
naïve.

The patch fixes it by adding a check to ensure that there are no more
jobs to be run.

Signed-off-by: Niels Thykier <niels@thykier.net>