c/fields: Retire no-source-field - it cannot be emitted
The tag "no-source-field" cannot be emitted by Lintian due to
dpkg-source/Dpkg::Source::Package, which refuses to extract a source
package missing (among others) the "Source" field.  This causes
collection/unpacked to fail and Lintian to skip the check of the
source package.

The tag cannot even be emitted via "-C fields" as checks/fields needs
"debfiles" (which depends on "unpacked").

Signed-off-by: Niels Thykier <niels@thykier.net>