L::Collect: Allow indirect "Needs-Info" dependencies
Instead of manually keeping track of transitive dependencies, allow
methods in L::Collect to declare a "symbolic" dependency on another
method by prefixing the dependency with a colon.

As an example, is_conffile (which relies entirely on the control
method) now declares a dependency on ":control" rather than
"bin-pkg-control".

The indirect dependencies are only allowed in L::Collect (and not in
the Needs-Info of collections or checks).

needs-info-missing.t has been updated to cope with these symbolic
dependencies and even "or'ed" dependencies, which were previous
silently ignored.  "or'ed" dependencies currently only happens when
the Source and the Binary variant of a method has different
dependencies (e.g. changelog).

Signed-off-by: Niels Thykier <niels@thykier.net>