L::Relation: Explicitly handle <pkg>:<arch> cases
Extract and check the <arch> part of a relation.  The Multi-Arch
spec[1] recommends assuming that if <arch> is present and not "any"
that the relation is unsatisfiable.
  Beyond that it is vague on how the <arch> part is applied in
practise (except that the package in the relation must be "Multi-Arch:
allowed").

This implementation assumes that if <arch> is present", the (only)
identity holds.  i.e. "pkg:a" implies "pkg:b" iff a == b.

This assumption is handy for identity relations and for finding
duplicated relations, where the check relies on implication to find
possible duplications.  However, it fails the cases where ":X" implies
":any" (for any valid value of X).

[1] https://wiki.ubuntu.com/MultiarchSpec

Signed-off-by: Niels Thykier <niels@thykier.net>