L::Unpacker: Reduce unpack jobs for new entries
The unpacker by default runs all collections on new entries, but this
is way too much with (e.g.) "-o -C control-files", where a simple
"bin-pkg-control" would suffice.

To solve this, strip out all nodes of the (cloned) collmap that will
never be needed.  It is not a perfect solution, but it reasonable in
practise.

Signed-off-by: Niels Thykier <niels@thykier.net>