L::Unpacker: Do not unpack unneeded colls even for new entries
For new entries, use the L::Profile to determine what collections are
needed during unpacking.  This means that e.g. "-oC files some.dsc" no
longer unpacks anything.  This is mostly useful for using -C/-X on a
.changes file.

This also fixes a regression of running too many collections with -F,
which was introduced at some time in the 2.5.11 development series.
2.5.10.X is not affected by the regression.

Signed-off-by: Niels Thykier <niels@thykier.net>