L::Profile: Turn "profile-path" into "include-path"
Generalize the profile-path member into an include-path that lists
"Lintian roots" rather than just "profile dirs".  As a consequence,
there is no longer a "single root" for a Profile and accordingly
the "root" member is removed.

Two important side-effects are that Lintian::Data can now query and
use the include-path and checks can now be loaded from
$ENV{'HOME'}/.lintian/checks or even /etc/lintian/checks.

Signed-off-by: Niels Thykier <niels@thykier.net>