L::Lab: Add new entries to the index immediately
When a caller requests a new entry, attempt to add it immediately the
the index as a transient entry.  This makes it possible to avoid a
call to get_deb_info for entries created via reporting/harness.

If it cannot be added to the index, then re-try when/if the caller
actually creates the entry.

Signed-off-by: Niels Thykier <niels@thykier.net>