L::Command: Work around memory leak in IPC::Run
By default, spawn will set the child's "STDIN" to ("<", \undef).
Unfortunately, SCALAR refs are leaked (even the undef case) so this
translates to a "leak by default".

The work around is to translate ("<", \undef) to ("<&-"), which
doesn't leak.  Though, this only fixes the default case, so the leak
is only fixed as long as SCALAR ref remain unused in Lintian.

According to [1] about 2KB are lost per leak.  Ignoring the
"short-lived" collections, this means a 4KB leak per deb-format check
and a leak for every "delete_dir" (Lab and Lab entry removals).

[1] https://rt.cpan.org/Public/Bug/Display.html?id=13660

Debian bug: #301774

Signed-off-by: Niels Thykier <niels@thykier.net>