coll/index: Use find to generate source index
Make coll/index depend on unpacked and use find to generate the index
of the unpacked source package.  This gives an accurate view of what
is in unpacked (particular the debian/ is now always present).

The previous indexing method has been moved to a new collection
src-orig-index, which only concerns itself with creating an index of
what is the "orig" tarballs.

Signed-off-by: Niels Thykier <niels@thykier.net>