L::L::Entry: Fix _init croaking on existing entries
In some cases, _init would croak on loading an entry that exists but
where the "package symlink" (i.e. the one pointing the underlying
package) is broken.  This happens whenever the Laboratory outlives the
package itself (e.g. on lintian.d.o).

Secondly, as both constructors now (almost) always have a path the
package, there is no need to do the "resolve symlink" trick to find
the "correct path".

Signed-off-by: Niels Thykier <niels@thykier.net>