L::Profile: Allow "Load-Checks" field in profiles
Add a new optional field to profiles called "Load-Checks".  It ensures
that a given list of checks are loaded to allow Enable-Tags or
Disable-Tags to work on third-party checks.

If a check is loaded via Load-Checks, all of its tags will be
disabled.  If a check listed in a Load-Checks field is already loaded
(by the time the field is processed), the request is silently ignored.

Signed-off-by: Niels Thykier <niels@thykier.net>