L::Profile: Always load the "lintian" check
The frontend and L::Tags will generally always attempt to emit the
tags of the "lintian" check.  When the check is not loaded, it cause
Lintian to abort with an error.

This patch ensures that L::Profile will load the "lintian" check in
profiles that did not load itself.  Furthermore, it will enable its
tags by default.

The main rationale behind (auto) enabling these tags is to assist
users in debugging their overrides files.  Without these tags enabled,
incorrect overrides will be "silently ignored".

For profiles that really do not want these tags, the check can be
disabled via a "Disable-Tags-From-Check".  This is currently done
in the ftp-master-auto-reject profile.

Signed-off-by: Niels Thykier <niels@thykier.net>