L::Processable: Properly handle "source-version"
When creating L::Lab::Entry instances, we currently rely on
L::Processable's new_from_metadata to init the L::Lab::Entry instance.
Unfortunately it previously discarded the "source-version" and relied
on extracting it from the "source" field.

The result is that the source-version field was effectively replaced
with the package's own version (as is the default).  This error could
then propagate back into the lab indices if these were changed in the
run.

Temporary labs are unaffected (it doesn't use indices) and static labs
can be "repaired" by a simple "remove and re-add" of affected
packages.

Signed-off-by: Niels Thykier <niels@thykier.net>