L::C::Simple: Fix wait method and use O(1) reap
The method variant of wait made it difficult to distinguish between
reaping a child that exited with 0 and returning from a non-blocking
wait.  The latter now returns -1 (like an error) but sets $! to 0.

Remove support for using L::C::Simple::wait as a function with no
arguments.  This case now returns -1 and sets $! to ENVAL.

When reaping children from a table (e.g. as done in L::Unpacker),
require the PID to be the key of the table.  This allows wait to do a
hash look up (average O(1)) instead of a linear scan over the table to
find the right element.

Signed-off-by: Niels Thykier <niels@thykier.net>