lintian: Set IPCRUNDEBUG to none
When loading IPC::Run in frontend/lintian, set IPCRUNDEBUG to "none"
to disable IPC::Run debugging.  We cannot rely on L::Command to do it
for us as _guess_version is run way too early.

At the same time, add an "unless exists" clause to the setting of
IPCRUNDEBUG.  This allows anyone interested to get the IPC::Run debug
information.

Signed-off-by: Niels Thykier <niels@thykier.net>