Add new collection interface to reduce overhead
The major overhead in running collections are loading the required
Perl Modules (again).  As an example, loading Lintian::Util has a cost
of 0.040 seconds (the first time).  With a couple of extra modules,
coll/hardening-info has a average "no-op" runtime of at least 0.150s.
Eliminating the module load overhead, it drops to 0.065s.

This patch set creates a new (backwards compatible) interface for
invoking collections.  Instead of doing a complete fork+exec, the new
interface loads the collection script like a perl module and load a
"collect" sub (similar to how scripts work).  This interface can still
be combined with fork for parallelization.

Because it avoids the exec, the child process retrains the cache of
(most of) the Perl Modules.  This vastly reduces the "start up" time
for the collections.

Signed-off-by: Niels Thykier <niels@thykier.net>