L::Profile: Allow and subst {VENDOR} in the profile name
Allow {VENDOR} to appear in the profile name - it can be used with the
"--profile" (or LINTIAN_PROFILE) to request the "most specific vendor
specific" for a profile.  As an example, Lintian's default choice for
a profile can now be described via:

  --profile {VENDOR}/main

With this, it is possible to create application specific profiles that
have vendor specific cases (without having to do the vendor mapping
manually).

Signed-off-by: Niels Thykier <niels@thykier.net>